Rename existing export.ts legacy_export.ts to free up file name
