+++ libs/backend/src/scan/cast_vote_records/legacy_export.test.ts
@@ -30,7 +30,7 @@ import {
   getCastVoteRecordReportStream,
   InvalidSheetFoundError,
   ResultSheet,
-} from './export';
+} from './legacy_export';
 
 const electionDefinition: ElectionDefinition = {
   ...electionMinimalExhaustiveSampleDefinition,
